part5: exercise 5.28 fix e2e tests for the routed app

+++ part5/bloglist-frontend/src/App.jsx
@@ -96,6 +96,10 @@ const App = () => {
   }
 
   const handleDelete = async (blog) => {
+    if (!window.confirm('Remove blog ' + blog.title + ' by ' + blog.author)) {
+      return
+    }
+
     try {
       await blogService.remove(blog.id)
